Added AuthContext and more functionality to the modal component
Developed auth context so user information can be accessed throughout the app. Modal component now has adjustable height (should be responsive in the future) and added button functionality. Also added signout button in Home to sign out. Also made it so that the default is home screen if there is a user.